#5815: Improve comparison function to better handle inf values

+++ tests/tt_eager/python_api_testing/non_working_unit_tests/grayskull/test_eltwise_recip.py
@@ -18,7 +18,7 @@ def run_eltwise_recip_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_
     if in_mem_config == "SYSTEM_MEMORY":
         in_mem_config = None
 
-    x = torch.Tensor(size=input_shape).uniform_(-100, 100)
+    x = torch.Tensor(size=input_shape).uniform_(-100, 100).to(torch.bfloat16)
     x_ref = x.detach().clone()
 
     # get ref result
